refactor: ディレクトリ構造をFeature-firstに合わせて修正
