+++ apps/dashboard/lib/core/ui/app.dart
