+++ apps/dashboard/lib/features/auth/ui/invitation_code_screen.dart
