+++ apps/dashboard/lib/features/event/ui/event_info_screen.dart
