+++ apps/dashboard/lib/features/sponsor/ui/sponsor_detail_screen.dart
