+++ apps/dashboard/lib/main.dart
@@ -1,7 +1,7 @@
 import 'dart:developer';
 
-import 'package:dashboard/app.dart';
 import 'package:dashboard/core/provider/environment.dart';
+import 'package:dashboard/core/ui/app.dart';
 import 'package:dashboard/features/auth/data/notifier/auth_notifier.dart';
 import 'package:dashboard/features/auth/data/provider/auth_service.dart';
 import 'package:flutter/foundation.dart';
